Relocate math types
Relocated to 'math' directory; Vector4 has been renamed to Color to better match its usage (Vector4 = x,y,z,w while this was r,g,b,a);
